+++ packages/core/src/odata/common/request-builder/batch/batch-request-options.ts
@@ -1,0 +1,13 @@
+import { Destination } from '../../../../scp-cf';
+
+export interface BatchRequestOptions {
+  subrequestPath: 'absolute' | 'servicePathRelative' | 'entityRelative';
+}
+
+export type BatchRequestSerializationOptions = BatchRequestOptions & {
+  destination?: Destination;
+};
+
+export const defaultOptions: BatchRequestOptions = {
+  subrequestPath: 'servicePathRelative'
+};