+++ packages/core/src/odata/common/request/odata-batch-request-config.ts
@@ -1,4 +1,8 @@
 import { v4 as uuid } from 'uuid';
+import {
+  BatchRequestOptions,
+  defaultOptions
+} from '../request-builder/batch/batch-request-options';
 import { ODataRequestConfig } from './odata-request-config';
 
 export class ODataBatchRequestConfig extends ODataRequestConfig {
@@ -6,6 +10,8 @@ export class ODataBatchRequestConfig extends ODataRequestConfig {
    * @deprecated Since v1.30.0.
    */
   static readonly content_type_prefix = 'multipart/mixed; boundary=batch_';
+
+  options: BatchRequestOptions = defaultOptions;
 
   /**
    * @deprecated Since v1.30.0. Use [[boundary]] instead.
@@ -28,6 +34,10 @@ export class ODataBatchRequestConfig extends ODataRequestConfig {
       'content-type': `multipart/mixed; boundary=${boundary}`,
       accept: 'multipart/mixed'
     });
+  }
+
+  withOptions(options: BatchRequestOptions) {
+    this.options = { ...this.options, ...options };
   }
 
   resourcePath(): string {