+++ packages/core/src/odata/v2/index.ts
@@ -3,7 +3,7 @@ export * from './entity';
 export * from './entity-deserializer';
 export * from './entity-serializer';
 export * from './batch-change-set';
-export * from './batch-request-builder';
+export * from './request-builder/batch-request-builder';
 export * from './batch-request-serializer';
 export * from './payload-value-converter';
 export * from './request-builder';