+++ packages/core/src/odata/v4/batch-request-builder.ts
@@ -10,6 +10,7 @@ import { BatchResponse } from '../common';
 import { parseBatchResponse } from '../common/request-builder/batch/batch-response-parser';
 import { BatchRequestBuilder } from '../common/request-builder/batch/batch-request-builder';
 import { deserializeBatchResponse } from '../common/request-builder/batch/batch-response-deserializer';
+import { serializeBatchRequest } from '../common/request-builder/batch/batch-request-serializer';
 import { responseDataAccessorV4 } from './request-builder/response-data-accessor';
 import { entityDeserializerV4 } from './entity-deserializer';
 
@@ -30,6 +31,13 @@ export class ODataBatchRequestBuilderV4 extends BatchRequestBuilder {
     options?: DestinationOptions
   ): Promise<BatchResponse[]> {
     return this.build(destination, options)
+      .then(request => {
+        this.requestConfig.payload = serializeBatchRequest(this, {
+          ...this.requestConfig.options,
+          destination: request.destination!
+        });
+        return request;
+      })
       .then(request => request.execute())
       .then(response => parseBatchResponse(response))
       .then(parsedResponse =>